% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catch_coverage.R
\name{Catch_coverage}
\alias{Catch_coverage}
\title{Catch_cov: function to check the coverage in Catch table}
\usage{
Catch_coverage(Catch_tab, SP, MS, GSA)
}
\arguments{
\item{Catch_tab}{Catch table in MED&BS format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}
}
\value{
summary table and plots
}
\description{
Catch_cov: function to check the coverage in Catch table
}
\examples{
Catch_coverage(Catch_tab_example,"DPS","ITA","9")
}
