% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catch_tab_example.R
\docType{data}
\name{Catch_tab_example}
\alias{Catch_tab_example}
\title{Catch table in MED&BS datacall format}
\format{
An object of class \code{data.frame} with 57 rows and 171 columns.
}
\usage{
Catch_tab_example
}
\description{
Catch table in MED&BS datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{MED&BS}
\keyword{datacall}
