% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_empty_field_VII2.R
\name{check_EF_taskVII2}
\alias{check_EF_taskVII2}
\title{Check empty fields in GFCM Task VII.2 table}
\usage{
check_EF_taskVII2(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.2 table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
Two lists are returned by the function. The first list gives the number of NA for each reference column. The second list gives the index of each NA in the reference column.
}
\description{
The function checks the presence of not allowed empty data in the given table, according to the \href{https://gfcm.sharepoint.com/sites/DCRF/Shared\%20Documents/Forms/AllItems.aspx?id=\%2Fsites\%2FDCRF\%2FShared\%20Documents\%2FGFCM\%2DDCRF\%2Dmanual\%2D2018\%2Dv\%2E21\%2E2\%2Epdf&parent=\%2Fsites\%2FDCRF\%2FShared\%20Documents&p=true&ga=1}{GFCM, 2018. GFCM Data Collection Reference Framework (DCRF). Version: 20.1}
}
\examples{
check_EF_taskVII2(task_vii2)

}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
