% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_empty_field_VII32.R
\name{check_EF_TaskVII32}
\alias{check_EF_TaskVII32}
\title{Check empty fields in GFCM Task VII.3.2 table}
\usage{
check_EF_TaskVII32(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.3.2 table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
two lists. The first list gives the number of NA for each column of interest. The second list gives the index of each NA in the column of interest
}
\description{
Check empty fields in GFCM Task VII.3.2 table
}
\examples{
check_EF_TaskVII32(task_vii32)

}
