% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_FDI_coverage.r
\name{check_FDI_coverage}
\alias{check_FDI_coverage}
\title{•	Coverage of data by table}
\usage{
check_FDI_coverage(data, MS, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
•	Coverage of data by table
}
\examples{
check_FDI_coverage(data=fdi_h_spatial_landings,MS="ITA", verbose = FALSE)
}
