% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_L50_VII31.R
\name{check_l50_TaskVII.3.1}
\alias{check_l50_TaskVII.3.1}
\title{Consistency of L50 values in Task VII.3.1 table}
\usage{
check_l50_TaskVII.3.1(data, MS, GSA, SP = NA, verbose = FALSE)
}
\arguments{
\item{data}{GFCM Task VII.3.1 table}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a plot to visualize the L50 time series by species and sex for the selected MS and GSA
}
\description{
Function to verify the consistency of L50 reported in the TaskVII.3.1 table by mean of plot by GSA, species and sex
}
\examples{
check_l50_TaskVII.3.1(task_vii31, MS = "ITA", GSA = "19",SP="HKE")
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
