% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_lfd_VII2.R
\name{check_ldf_TaskVII.2}
\alias{check_ldf_TaskVII.2}
\title{Consistency of length frequency distributions}
\usage{
check_ldf_TaskVII.2(data, MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task II.2 table}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a table with the comparison between min/max lengths observed for each species with theoretical values.
}
\description{
Function to verify the consistency of the length frequency distributions (LFD) reported in the TaskVII.2 table. The functions generates plots of the LFD by species, source, segment and year.
}
\examples{
check_ldf_TaskVII.2(task_vii2, MS = "ITA", GSA = "18", SP = "HKE")
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
