% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lengths_MEDBS_AR.r
\name{check_lengths_MEDBS_AR}
\alias{check_lengths_MEDBS_AR}
\title{Cross check of number of lengths between MED&BS catch table and AR Table 2.1}
\usage{
check_lengths_MEDBS_AR(
  MEDBS,
  AR,
  MS,
  GSA,
  SP,
  year,
  species_list = SPs,
  OUT = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{MEDBS}{data frame containing MED&BS catch table}

\item{AR}{data frame containing Annual Report data}

\item{MS}{member state code}

\item{GSA}{Geographical Subarea}

\item{SP}{Species 3-alpha code}

\item{year}{Reference year for the analysis}

\item{species_list}{table of species 3-alpha codes, reporting the MED & BS mandatory species}

\item{OUT}{Default is FALSE. If set as TRUE tables in csv will be saved in the OUTPUT folder created in the working directory}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a data frame reporting the total number of length measurements related to catches, landings and discards conducted in the selected country and year. MED & BS data will be reported as sum of the measurements by year and sum of the trips reported by quarter in the year.
}
\description{
The function compares the number of length measurements reported in MED & BS data call catch table with the information reported in the table 2.1 ("Stocks") of the Annual Report.
The user have to define the two source tables as data frames in the \code{MEDBS} and \code{AR} parameters. The analysis should be constrained at the selected country level (\code{MS}) and at the selected \code{year}.
}
\author{
Walter Zupa \href{mailto:zupa@fondazionecoispa.org}{zupa@fondazionecoispa.org}
}
