% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_LFD_comm_cat.R
\name{check_LFD_comm_cat}
\alias{check_LFD_comm_cat}
\title{Check consistency of LFD by year and commercial category}
\usage{
check_LFD_comm_cat(data_example)
}
\arguments{
\item{data_example}{RCG CS table}
}
\value{
plot and a summary table with ranges by year and commercial category
}
\description{
Check consistency of LFD by year and commercial category
}
\examples{
check_LFD_comm_cat(data_ex)
}
