% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_lmat_TaskVII32.R
\name{check_lmat_TaskVII.3.2}
\alias{check_lmat_TaskVII.3.2}
\title{Plot of the maturity stages per length for each sex and species}
\usage{
check_lmat_TaskVII.3.2(data)
}
\arguments{
\item{data}{GFCM Task VII.3.2 table}
}
\value{
Plot of the maturity stages per length for each sex and species
}
\description{
Plot of the maturity stages per length for each sex and species
}
\examples{
check_lmat_TaskVII.3.2(task_vii32)
}
