% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loc.R
\name{check_loc}
\alias{check_loc}
\title{Check trip location}
\usage{
check_loc(data_example)
}
\arguments{
\item{data_example}{detailed data in RCG CS format}
}
\value{
map of trip locations
}
\description{
If Initial and/or Final coordinates are included in the data, maps of them are produced. If not the locations of the harbours are mapped.
}
\examples{
check_loc(data_ex)
}
