% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_relatioship_length_weight_VII2.R
\name{check_lw_TaskVII.2}
\alias{check_lw_TaskVII.2}
\title{Plot of the relationship length weight for each species}
\usage{
check_lw_TaskVII.2(data, MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.2 table}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
The function return a plot of the length weight relationship per species.
}
\description{
Function to check the consistency of length-weight relationship in the GFCM Task VII.2 table per species.
}
\examples{
check_lw_TaskVII.2(task_vii2, MS = "ITA", GSA = "18", SP = "BOG")
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
