% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mat.R
\name{check_mat}
\alias{check_mat}
\title{Check consistency sex and maturity stage}
\usage{
check_mat(data_example, immature_stages = c("0", "1", "2a"))
}
\arguments{
\item{data_example}{table of detailed data in RCG format}

\item{immature_stages}{maturity stages considered immature}
}
\value{
Plot ogive by sex
}
\description{
Check consistency sex and maturity stage
}
\examples{
check_mat(data_ex,immature_stages=c("0","1","2a"))
}
