% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_min_max_L50_VII31.R
\name{check_minmaxl50_TaskVII.3.1}
\alias{check_minmaxl50_TaskVII.3.1}
\title{Comparison between min/max L50 observed for each species and sex with theoretical values}
\usage{
check_minmaxl50_TaskVII.3.1(data1, data2)
}
\arguments{
\item{data1}{GFCM Task VII.3.1 table}

\item{data2}{Theoretical values of min/max L50 for each species and sex}
}
\value{
The function returns a table with the comparison between min/max L50 observed for each species and sex with theoretical values.
}
\description{
Function to verify the consistency of L50 reported in the TaskVII.3.1 table with the theoretical values reported in the minmaxLtaskVII31 table. The function allows to identify the records in which the observed L50 are greater or lower than the expected ones.
}
\examples{
check_minmaxl50_TaskVII.3.1(task_vii31,minmaxLtaskVII31)
}
