% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_min_max_lenght_VII2.R
\name{check_minmaxl_TaskVII.2}
\alias{check_minmaxl_TaskVII.2}
\title{Comparison between min/max observed for each species with theoretical values}
\usage{
check_minmaxl_TaskVII.2(data1, data2)
}
\arguments{
\item{data1}{GFCM Task II.2 table}

\item{data2}{Theoretical values of min/max for each species}
}
\value{
Table for comparison between min/max observed for each species with theoretical values
}
\description{
Comparison between min/max observed for each species with theoretical values
}
\examples{
check_minmaxl_TaskVII.2(task_vii2,minmaxLtaskVII2)
}
