% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_trips_MEDBS_AR.r
\name{check_n_trips_MEDBS_AR}
\alias{check_n_trips_MEDBS_AR}
\title{Cross check of number of trips between MED&BS catch table and AR Table 2.5}
\usage{
check_n_trips_MEDBS_AR(MEDBS, AR, MS, year, OUT = FALSE, verbose = FALSE)
}
\arguments{
\item{MEDBS}{data frame containing MED&BS catch table}

\item{AR}{data frame containing Annual Report data}

\item{MS}{member state code}

\item{year}{Reference year for the analysis}

\item{OUT}{Default is FALSE. If set as TRUE tables in csv will be saved in the OUTPUT folder created in the working directory}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a data frame reporting the total number of trips related to catches, landings and discards conducted in the selected country and year. MED & BS data will be reported both as sum of the trips by year and as sum of the trips reported by quarter in the year.
}
\description{
The function compares the number of trips reported in MED & BS data call catch table with the information reported in the table 2.5 ("Sampling plan biol") of the Annual Report.
The user have to define the two source tables as data frames in the \code{MEDBS} and \code{AR} parameters. The analysis should be constrained at the selected country level (\code{MS}) and at the selected \code{year}.
}
\author{
Walter Zupa \href{mailto:zupa@fondazionecoispa.org}{zupa@fondazionecoispa.org}
}
