% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_presence_GSA_FS_II2.R
\name{check_presence_taskII2}
\alias{check_presence_taskII2}
\title{Check of missing combination GSA/Fleet segment per year}
\usage{
check_presence_taskII2(data1, data2)
}
\arguments{
\item{data1}{GFCM Task II.2 table}

\item{data2}{Combination for Task II.2 table}
}
\value{
List of missing combination GSA/Fleet segment per year
}
\description{
Check of missing combination GSA/Fleet segment per year
}
\examples{
check_presence_taskII2(task_ii2,combination_taskII2)

}
