% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_record_duplicated_A.R
\name{check_RD_FDI_A}
\alias{check_RD_FDI_A}
\title{Check duplicated records in FDI A table}
\usage{
check_RD_FDI_A(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task A table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns the indices of the duplicated rows, checking the unique combinations of the first 19 columns of the FDI A table.
}
\description{
The function check the presence of duplicated records. In particular, it checks whether the combination of the first 19 columns generates duplicate records.
}
\examples{
check_RD_FDI_A(fdi_a_catch)
}
