% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_record_duplicated_G.R
\name{check_RD_FDI_G}
\alias{check_RD_FDI_G}
\title{Check duplicated records in FDI G table}
\usage{
check_RD_FDI_G(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task G table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 15 columns
}
\description{
Check duplicated records in FDI G table
}
\examples{
check_RD_FDI_G(fdi_g_effort)
}
