% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_record_duplicated_H.R
\name{check_RD_FDI_H}
\alias{check_RD_FDI_H}
\title{Check duplicated records in FDI H table}
\usage{
check_RD_FDI_H(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task H table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 15 columns
}
\description{
Check duplicated records in FDI H table
}
\examples{
check_RD_FDI_H(fdi_h_spatial_land)
}
