% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_record_duplicated_J.R
\name{check_RD_FDI_J}
\alias{check_RD_FDI_J}
\title{Check duplicated records in FDI J table}
\usage{
check_RD_FDI_J(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task J table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns the indices of the duplicated rows, checking the unique combinations of the first 7 columns of the FDI G table.
}
\description{
The function check the presence of duplicated records. In particular, it checks whether the combination of the first 15 columns generates duplicate records.
}
\examples{
check_RD_FDI_J(fdi_j_capacity)
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
