% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_record_duplicated_III.R
\name{check_RD_taskIII}
\alias{check_RD_taskIII}
\title{Check duplicated records in GFCM Task III table}
\usage{
check_RD_taskIII(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task III table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 10 columns
}
\description{
Check duplicated records in GFCM Task III table
}
\examples{
check_RD_taskIII(task_iii)
}
