% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_record_duplicated_VII2.R
\name{check_RD_taskVII2}
\alias{check_RD_taskVII2}
\title{Check duplicated records in GFCM Task VII.2 table}
\usage{
check_RD_taskVII2(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.2 table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 8 columns
}
\description{
Check duplicated records in GFCM Task VII.2 table
}
\examples{
check_RD_taskVII2(task_vii2)
}
