% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_record_duplicated_VII31.R
\name{check_RD_taskVII31}
\alias{check_RD_taskVII31}
\title{Check duplicated records in GFCM Task VII.3.1 table}
\usage{
check_RD_taskVII31(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.3.1 table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns the indices of the duplicated rows, checking the unique combinations of the first 5 columns of the Task Task VII.3.1 table.
}
\description{
The function check the presence of duplicated records. In particular, it checks whether the combination of the first 5 columns generates duplicate records.
}
\examples{
check_RD_taskVII31(task_vii31)
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
