% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_check_species_catfau_VII32.R
\name{check_species_catfau_TaskVII.3.2}
\alias{check_species_catfau_TaskVII.3.2}
\title{Check mismatching species/Catfau and Sex per maturity stages for Task VII.3.2 table}
\usage{
check_species_catfau_TaskVII.3.2(data, species, matsex)
}
\arguments{
\item{data}{GFCM Task VII.3.2 table}

\item{species}{List of combination of species/faunistic category for Task VII.3.2 table}

\item{matsex}{List of combination of sex/maturity stages for Task VII.3.2 table}
}
\value{
Two lists are returned by the function. The first output gives the list of mismatching combination of species/faunistic categories. The second output gives the list of mismatching combination of sex/maturity stages.
}
\description{
Function to check the correct codification of faunistic category according to species and sex in Task VII.3.2 table.
}
\examples{
check_species_catfau_TaskVII.3.2(task_vii32,catfau_check,sex_mat)
}
