% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKLANDGEARQUARTER_RDBFIS.R
\name{Check_Tot_Land_gear_Q}
\alias{Check_Tot_Land_gear_Q}
\title{Comparing total landings in weight between MEDBS and FDI EU Data Calls by Country, GSA, Species and Gear}
\usage{
Check_Tot_Land_gear_Q(
  data,
  data1,
  MS,
  GSA,
  SP,
  MEDBSSP,
  verbose = TRUE,
  OUT = FALSE
)
}
\arguments{
\item{data}{data frame containing MEDBS landings by length table}

\item{data1}{data frame containing FDI Table A Catch data}

\item{MS}{member state 3 alpha code}

\item{GSA}{FAO Geographical Sub Area code (GSA9, GSA29, etc)}

\item{SP}{ASFIS FAO 3alpha code of the choosen species}

\item{MEDBSSP}{look up table in which requested MEDBS Data Call species (see MEDBS data call Annex I) are listed}

\item{verbose}{boolean. If TRUE a message is printed}

\item{OUT}{Default is FALSE. If set as TRUE plots and tables in csv will be saved in the OUTPUT folder created in the working directory}
}
\value{
The function returns plots comparison by year of the landings in weight provided through the MEDBS and FDI Data Calls. Moreover a csv file with the actual values will be created. Plot, csv and rds output files will be saved in the dedicated folders in OUTPUT folder
}
\description{
The function compares the landings in weight values reported in the MEDBS landings by length table with the one reported in the FDI Table A catch table. The analysis is done at MS, GSA, species and gear level
}
\examples{
\donttest{
# Check_Tot_Land_gear_Q (MEDBS,FDI,"ITA","GSA10","HKE",MEDBSSP)
}
# The function works by one country, subarea and species each.
# It is not possible assign more country, subarea or species.
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@irbim.cnr.it}{alessandro.mannini@irbim.cnr.it}
}
