% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circabc.R
\docType{data}
\name{circabc}
\alias{circabc}
\title{Ports coordinate according to codification CIRCABC}
\format{
An object of class \code{data.frame} with 24335 rows and 5 columns.
}
\usage{
circabc
}
\description{
Ports coordinate according to codification CIRCABC
}
\author{
CIRCABC \email{https://circabc.europa.eu/faces/jsp/extension/wai/navigation/container.jsp}
}
\keyword{CIRCABC}
\keyword{Harbour}
