% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ex}
\alias{data_ex}
\title{RCG CS example}
\format{
An object of class \code{data.frame} with 9099 rows and 40 columns.
}
\usage{
data_ex
}
\description{
RCG CS example
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{CS}
\keyword{RCG}
\keyword{samplingData}
