% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWG_length_ind.r
\name{EWG_length_ind}
\alias{EWG_length_ind}
\title{Main length size indicators}
\usage{
EWG_length_ind(
  data,
  type,
  out = "mean",
  MS,
  GSA,
  SP,
  splines = c(0.2, 0.4, 0.6, 0.8),
  Xtresholds = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{data}{data frame of landings or discards data}

\item{type}{type of data frame. "l" for landing and "d" for discard}

\item{out}{type of oututplot to be generated. "mean" or "median"}

\item{MS}{member state code as it is reported in the landing data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{splines}{spline values assignment to fit cumulative distributions}

\item{Xtresholds}{threshold value}
}
\description{
Main length size indicators
}
\examples{
EWG_length_ind(landing,type="l",out="mean",MS=c("ITA"),
GSA=c("18"),SP="NEP", splines=c(0.2,0.4,0.6,0.8),
Xtresholds = c(0.25,0.5,0.75))
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
