% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWG_plot_disc_vol.r
\name{EWG_plot_disc_vol}
\alias{EWG_plot_disc_vol}
\title{Plot of total discards by gear and fishery}
\usage{
EWG_plot_disc_vol(data, MS, GSA, SP)
}
\arguments{
\item{data}{data frame containing discard data}

\item{MS}{member state code as it is reported in the discard data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}
}
\value{
The function returns a plot of the total discards time series by fishery and gear
}
\description{
Plot of total discards by gear and fishery
}
\examples{
EWG_plot_disc_vol(data=discards,MS="ITA",GSA=18,SP="MUT")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
