% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWG_weight_minus1.R
\name{EWG_weight_minus1}
\alias{EWG_weight_minus1}
\title{weight -1 in landings}
\usage{
EWG_weight_minus1(data, type = "l", MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{data.table object containing landing or discards data}

\item{type}{type of table: "l" for landings; "d" for discards}

\item{MS}{member state code as it is reported in the landing data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{Boolean value to obtain further explanation messages from the function}
}
\value{
The function returns the number of rows with -1 values in landing weights having length class filled in.
}
\description{
The function checks landings in weight equal to -1 having length class filled in
}
\examples{
EWG_weight_minus1(data=landing,type="l",MS="ITA",GSA=18,SP="HKE",verbose=TRUE)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
