% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{fdi_a_catch}
\alias{fdi_a_catch}
\title{fdi_a_catch in FDI DGMAREMED&BS datacall format}
\format{
An object of class \code{data.frame} with 2600 rows and 23 columns.
}
\usage{
fdi_a_catch
}
\description{
fdi_a_catch in FDI DGMAREMED&BS datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{DGMAREMED&BS}
\keyword{FDI}
\keyword{datacall}
