% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_checks_spatial_HI.R
\name{FDI_checks_spatial_HI}
\alias{FDI_checks_spatial_HI}
\title{check NA values in spatial columns of both table H and I}
\usage{
FDI_checks_spatial_HI(data, MS, verbose = FALSE)
}
\arguments{
\item{data}{data frame of one between FDI table H (Landings by rectangle) and table I (Effort by rectangle)}

\item{MS}{Country code}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list of 2 tables. The first (NA_inconsistencies) reports the records with the inconsistencies detected in spatial of the selected table, while the second one reports the record without any sub_region assignment.
}
\description{
The function checks the  incorrect combination of NA in the spatial columns in both table H (Landings by rectangle) and table I (Effort by rectangle). The following check are included:
\enumerate{
\item{presence of NA values in 'c_square' field when 'rectangle_type', 'rectangle_lat', 'rectangle_lon' are all NA;}
\item{the presence any data in 'rectangle_type', 'rectangle_lat', 'rectangle_lon' when 'c_square' is reported;}
\item{the presence of any data in 'c-squares' when 'rectangle_type', 'rectangle_lat', 'rectangle_lon' are filled in.}
}
Furthermore the function identifies the records without any sub-region assignment.
}
\examples{
FDI_checks_spatial_HI(data = fdi_h_spatial_landings, MS = "PSP", verbose = TRUE)
FDI_checks_spatial_HI(data = fdi_i_spatial_effort, MS = "PSP", verbose = TRUE)
}
\author{
Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
