% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_cov_tableJ.R
\name{FDI_cov_tableJ}
\alias{FDI_cov_tableJ}
\title{Check number of record in FDI J table}
\usage{
FDI_cov_tableJ(
  data,
  MS,
  GSA,
  vessel_len = "COMBINED",
  fishtech = "COMBINED",
  verbose = TRUE
)
}
\arguments{
\item{data}{FDI table J capacity}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{vessel_len}{vessels length code ("COMBINED" values perform the analysis for all vessels length present in data)}

\item{fishtech}{selected fishing techniques ("COMBINED" values perform the analysis for all fishing techniques present in data)}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list. The first element gives the summary table of records number. From the second to the fifth element gives 4 plots for each variables among:
\itemize{
\item total trips;
\item total kW;
\item total GT;
\item total vessels;
}
}
\description{
The function checks and count the numbers of records data in the given table grouped by year, GSA, MS, vessels length, and fishing techniques for the following 4 variables: total trips; total kW; total GT; total vessels.
}
\examples{
FDI_cov_tableJ(data = fdi_j_capacity, MS = "PSP", GSA = "GSA99")
FDI_cov_tableJ(data = fdi_j_capacity, MS = "PSP", GSA = "GSA99", fishtech = c("DTS", "PGP"))
}
