% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_coverage.r
\name{FDI_coverage}
\alias{FDI_coverage}
\title{Coverage of data by table}
\usage{
FDI_coverage(data, MS, verbose = TRUE)
}
\arguments{
\item{data}{data frame. One of the allowed FDI table among A, G, H, I, J}

\item{MS}{Country}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
the function returns a data frame reporting the coverage of the selected table in terms of number of records by country, GSA and year
}
\description{
Coverage of data by table
}
\examples{
FDI_coverage(data=fdi_h_spatial_landings,MS="ITA", verbose = FALSE)
}
