% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Cross_checks_AG.R
\name{FDI_cross_checks_AG}
\alias{FDI_cross_checks_AG}
\title{Cross check between FDI tables A and G}
\usage{
FDI_cross_checks_AG(data1, data2, verbose = FALSE)
}
\arguments{
\item{data1}{FDI table A catch}

\item{data2}{FDI table G effort}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a table where all the miss matches between landings and effort are shown.
}
\description{
The function checks the possible data inconsistency between landings and effort.
}
\examples{
FDI_cross_checks_AG(data1 = fdi_a_catch, data2 = fdi_g_effort)
FDI_cross_checks_AG(fdi_a_catch, fdi_g_effort)
}
