% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Cross_Checks_IG.R
\name{FDI_cross_checks_IG}
\alias{FDI_cross_checks_IG}
\title{Cross check between FDI tables I and G}
\usage{
FDI_cross_checks_IG(data1, data2, verbose = FALSE)
}
\arguments{
\item{data1}{FDI spatial effort in table I}

\item{data2}{FDI effort table G}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list with two tables. In the first table all the mismatches between spatial effort in table I and effort in table G are shown, in the second table the comparison between total spatial effort of table I and total effort in table G is shown.
}
\description{
The function checks the possible data inconsistency between spatial effort in table I and effort in table G.
}
\examples{
FDI_cross_checks_IG(data1 = fdi_i_spatial_effort, data2 = fdi_g_effort)
}
\author{
Andrea Pierucci \href{mailto:pierucci@coispa.eu}{pierucci@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
