% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_disc_coverage.r
\name{FDI_disc_coverage}
\alias{FDI_disc_coverage}
\title{Coverage of FDI discard data}
\usage{
FDI_disc_coverage(data, MS, verbose = TRUE)
}
\arguments{
\item{data}{data frame of the FDI table A}

\item{MS}{Country}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list of data frames by GSA reporting the landing volumes (with discard >0, =0 and =NK and total landing) by year
}
\description{
The functions checks the discard coverage in table A for the selected MS by GSAs.
}
\examples{
FDI_disc_coverage(fdi_a_catch, MS = "PSP", verbose = TRUE)
}
