% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_fishdays_cov.R
\name{FDI_fishdays_cov}
\alias{FDI_fishdays_cov}
\title{Coverage comparison of totfishdays between FDI tables G and I}
\usage{
FDI_fishdays_cov(dataG, dataI, MS, verbose = TRUE)
}
\arguments{
\item{dataG}{data frame of FDA table G}

\item{dataI}{data frame of FDA table I}

\item{MS}{country code}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a data frame of the comparison of totfishdays between FDI tables G and I
}
\description{
The function checks the coverage comparison of totfishdays in FDI tables G and I
}
\examples{
FDI_fishdays_cov (dataG=fdi_g_effort, dataI=fdi_i_spatial_effort, MS="PSP", verbose = TRUE)
}
