% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{fdi_h_spatial_landings}
\alias{fdi_h_spatial_landings}
\title{fdi_h_spatial_landings in FDI DGMAREMED&BS datacall format}
\format{
An object of class \code{data.frame} with 2450 rows and 18 columns.
}
\usage{
fdi_h_spatial_landings
}
\description{
fdi_h_spatial_landings in FDI DGMAREMED&BS datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{DGMAREMED&BS}
\keyword{FDI}
\keyword{datacall}
