% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{fdi_i_spatial_effort}
\alias{fdi_i_spatial_effort}
\title{fdi_i_spatial_effort in FDI DGMAREMED&BS datacall format}
\format{
An object of class \code{data.frame} with 2450 rows and 16 columns.
}
\usage{
fdi_i_spatial_effort
}
\description{
fdi_i_spatial_effort in FDI DGMAREMED&BS datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{DGMAREMED&BS}
\keyword{FDI}
\keyword{datacall}
