% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_prices_not_null.r
\name{FDI_prices_not_null}
\alias{FDI_prices_not_null}
\title{Check of species value}
\usage{
FDI_prices_not_null(data, MS, SP = NA, verbose = FALSE)
}
\arguments{
\item{data}{data frame of FDA table A catch.}

\item{MS}{country code}

\item{SP}{vector of the species code for which the check should be performed}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
the function returns a list of two data frames. The first one reports the prices comparison by species, while the second one reports the cases in which total landings > 0 but landings value = 0.
}
\description{
The function estimates from the FDI table A an average price per species and year and compares it with average price calculated per country (by species). Furthermore, the function performs comparisons between total weight landings and total value landings. In particular it identifies the cases with total landings > 0 but landings value = 0. In case SP parameter is not specified, the analysis is conducted over all the species in the provided data frame.
}
\examples{
FDI_prices_not_null(
  data = fdi_a_catch, MS = "PSP",
  SP = c("ARA", "BOG", "HKE"), verbose = TRUE
)
}
\author{
Vasiliki Sgardeli \href{mailto:vsgard@hcmr.gr}{vsgard@hcmr.gr}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
