% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_species_price.r
\name{FDI_species_price}
\alias{FDI_species_price}
\title{Check of species value}
\usage{
FDI_species_price(data, MS, SP = NA, make_plot = F, verbose = FALSE)
}
\arguments{
\item{data}{data frame of FDA table A catch.}

\item{MS}{country code}

\item{SP}{vector of the species code for which the check should be performed}

\item{make_plot}{boolean. If TRUE plots are returned.}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\description{
The function finds an average price per species and year and compare it to the average price calculated per country.
}
\examples{
FDI_species_price(data=fdi_a_catch, MS="PSP",
 SP=c("ARA","BOG","HKE"), make_plot=TRUE, verbose = TRUE)

 FDI_species_price(data=fdi_a_catch, MS="PSP",
 SP=c("ARS"), make_plot=TRUE, verbose = TRUE)
}
