% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_ts_tableG.R
\name{FDI_ts_tableG}
\alias{FDI_ts_tableG}
\title{Check empty fields in FDI G table}
\usage{
FDI_ts_tableG(
  data,
  MS,
  GSA,
  vessel_len = "COMBINED",
  fishtech = "COMBINED",
  met = "COMBINED",
  verbose = TRUE
)
}
\arguments{
\item{data}{FDI table G effort}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{vessel_len}{vessels length code ("COMBINED" values perform the analysis for all vessels length present in data)}

\item{fishtech}{selected fishing techniques ("COMBINED" values perform the analysis for all fishing techniques present in data)}

\item{met}{selected metiers ("COMBINED" values perform the analysis for all metiers present in data)}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list. The first element gives the summary table of records number. From the second to the nineth element gives 8 plots for each variables among:
\itemize{
\item totseadays,
\item totfishdays,
\item totkwdaysatsea,
\item totgtdaysatsea,
\item totkwfishdays,
\item totgtfishdays,
\item hrsea,
\item kwhrsea
}
}
\description{
The function checks and count the numbers of records data in the given table grouped by year, GSA, MS, vessels length, fishing techniques, and metier for the following 8 variables: Total days at sea; Total Fishing Days; Total kW days at Sea; total GT days at sea; Total kW fishing days; totgtfishdays; Hours at Sea; kW hours at sea;
according to the \href{https://datacollection.jrc.ec.europa.eu/documents/10213/1385040/FDI2021-annex.pdf/6bb0a9b7-166c-48c8-ad12-48ea59a29ffe}{Fisheries Dependent Information data call 2021 - Annex 1} If Vessel length, fishing technique, and metier are not specified by the user the function combines those by default.
}
\examples{
FDI_ts_tableG(data=fdi_g_effort, MS="PSP", GSA="GSA99")
FDI_ts_tableG(fdi_g_effort, MS="PSP", GSA="GSA99", fishtech="DTS", met="OTB_MDD_>=40_0_0")
}
