% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_ts_tableJ.R
\name{FDI_ts_tableJ}
\alias{FDI_ts_tableJ}
\title{Check empty fields in FDI J table}
\usage{
FDI_ts_tableJ(
  data,
  MS,
  GSA,
  vessel_len = "COMBINED",
  fishtech = "COMBINED",
  verbose = TRUE
)
}
\arguments{
\item{data}{FDI table J capacity}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{vessel_len}{vessels length code ("COMBINED" values perform the analysis for all vessels length present in data)}

\item{fishtech}{selected fishing techniques ("COMBINED" values perform the analysis for all fishing techniques present in data)}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list. The first element gives the summary table of records number. From the second to the fifth element gives 4 plots for each variables among:
\itemize{
\item total trips;
\item total kW;
\item total GT;
\item total vessels;
}
}
\description{
The function checks and count the numbers of records data in the given table grouped by year, GSA, MS, vessels length, and fishing techniques for the following 4 variables: total trips; total kW; total GT; total vessels;
according to the \href{https://datacollection.jrc.ec.europa.eu/documents/10213/1385040/FDI2021-annex.pdf/6bb0a9b7-166c-48c8-ad12-48ea59a29ffe}{Fisheries Dependent Information data call 2021 - Annex 1}
}
\examples{
FDI_ts_tableJ(data=fdi_j_capacity, MS="PSP", GSA="GSA99")
FDI_ts_tableJ(data=fdi_j_capacity, MS="PSP", GSA="GSA99", fishtech = c("DTS","PGP"))
}
