% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_vessel_lenth.r
\name{FDI_vessel_lenth}
\alias{FDI_vessel_lenth}
\title{Check of vessel lenght in FDI table J}
\usage{
FDI_vessel_lenth(data, MS, verbose = TRUE)
}
\arguments{
\item{data}{data frame of table J}

\item{MS}{country code}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\description{
The function checks the average length vessels with the vessel length category (table J)
}
\examples{
FDI_vessel_lenth(data=fdi_j_capacity, MS="PSP", verbose = TRUE)
}
