% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_vessel_numbers.r
\name{FDI_vessel_numbers}
\alias{FDI_vessel_numbers}
\title{Check number of vessels in FDI table J and G}
\usage{
FDI_vessel_numbers(dataJ, dataG, MS, verbose = TRUE)
}
\arguments{
\item{dataJ}{data frame of FDA table J catch.}

\item{dataG}{data frame of FDA table G catch.}

\item{MS}{country code}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list of data frames. The first element reports the number of vessel in table J in comparison with table G and the relative difference percentage, while the second one reports the vessels not present in table G.
}
\description{
The function cross-checks the number of vessels in table J in comparison with the number reported in table G.
}
\examples{
FDI_vessel_numbers(dataJ = fdi_j_capacity, dataG = fdi_g_effort, MS = "PSP", verbose = TRUE)
}
