% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_check_headers.r
\name{GFCM_check_headers}
\alias{GFCM_check_headers}
\title{Check headers of GFCM tables}
\usage{
GFCM_check_headers(data, table)
}
\arguments{
\item{data}{data frame of data for the reference table}

\item{table}{character, vector specifying the reference table. Allowed values: "task2.2", "task3", "task7.2", "task7.3.1", "task7.3.2"}
}
\value{
In case the data frame reports has the headers in the format of the RDBFIS database, it is returned back with the headers in the GFCM format adopted also by RDBqc package
}
\description{
Check headers of GFCM tables
}
