% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_cov_task3.R
\name{GFCM_cov_task3}
\alias{GFCM_cov_task3}
\title{Check number of individuals in GFCM Task III table}
\usage{
GFCM_cov_task3(data, MS, GSA, SP = "COMBINED", verbose = TRUE)
}
\arguments{
\item{data}{Task III table}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format ("COMBINED" values perform the analysis for all species present in data)}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a list. The first element is a table reporting the inconsistencies of the total individual caught in comparison to the sum of the individuals alive, dead and released. The second element of the list is a plot of the time series of the individuals caught by year, source, and gear.
}
\description{
The function checks the consistencies of the total number of individual caught in comparison with the sum of the individuals alive, dead and released reported in Task III table.
}
\examples{
GFCM_cov_task3(data = task_iii, SP = "Dasyatis pastinaca",
MS = "ITA", GSA = "18")
}
\author{
Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
