% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{GSAlist}
\alias{GSAlist}
\title{GSAlist table}
\format{
An object of class \code{data.frame} with 22 rows and 2 columns.
}
\usage{
GSAlist
}
\description{
GSAlist table
}
\author{
Walter Zupa \email{zupa@fondazionecoispa.org}
}
\keyword{GSAlist}
