% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{GSAs}
\alias{GSAs}
\title{GSAs table}
\format{
An object of class \code{data.frame} with 22 rows and 2 columns.
}
\usage{
GSAs
}
\description{
GSAs table
}
\author{
Walter Zupa \email{zupa@fondazionecoispa.org}
}
\keyword{GSAs}
