% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landing_coverage.R
\name{Landing_coverage}
\alias{Landing_coverage}
\title{Catch_cov: function to check the coverage in Catch table}
\usage{
Landing_coverage(Landing_tab, SP, MS, GSA)
}
\arguments{
\item{Landing_tab}{Landing table in MED&BS format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}
}
\value{
summary table and plots
}
\description{
Catch_cov: function to check the coverage in Catch table
}
\examples{
Landing_coverage(Landing_tab_example,"DPS","ITA","9")
}
