% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LW_check_MED_BS.R
\name{LW_check_MED_BS}
\alias{LW_check_MED_BS}
\title{LW params in GP_tab in table check#'}
\usage{
LW_check_MED_BS(GP_tab, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{GP_tab}{growth params table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a summary table and plots of the LW params
}
\description{
The function allows to check the length-weight parameters included in the GP table for a selected species
}
\examples{
LW_check_MED_BS(GP_tab_example,"MUT","ITA","18")
}
