% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_tab_check.R
\name{MA_tab_check}
\alias{MA_tab_check}
\title{MA_tab (maturity at age) table check}
\usage{
MA_tab_check(MA_tab, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{MA_tab}{maturity at AGE table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a summary table and plots
}
\description{
MA_tab (maturity at age) table check
}
\examples{
MA_tab_check(MA_tab_example,"DPS","ITA","9")
}
