% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_ALK_MLAA.r
\name{MEDBS_ALK_MLAA}
\alias{MEDBS_ALK_MLAA}
\title{Plot of Age-Length Keys}
\usage{
MEDBS_ALK_MLAA(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{ALK table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned.}
}
\value{
a list of ALK plots by sex is return.
}
\description{
The function allows to check the data in the ALK table providing plots of the mean length by age by sex for a selected species
}
\examples{
MEDBS_ALK_MLAA(data = ALK_tab_example, SP = "MUT", MS = "ITA", GSA = "GSA 99")
}
\author{
Isabella Bitetto \href{mailto:bitetto@fondazionecoispa.org}{bitetto@fondazionecoispa.org}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
