% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_ALK_NB.r
\name{MEDBS_ALK_NB}
\alias{MEDBS_ALK_NB}
\title{Check on numbers in ALK}
\usage{
MEDBS_ALK_NB(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{ALK table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned.}
}
\value{
a list of recrods with mismatches is return.
}
\description{
The function allows to verify the numbers of individuals in the ALK table by sex and age for a selected species
}
\examples{
MEDBS_ALK_NB(data = ALK_tab_example, SP = "MUT", MS = "ITA", GSA = "GSA 99")
}
\author{
Isabella Bitetto \href{mailto:bitetto@fondazionecoispa.org}{bitetto@fondazionecoispa.org}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
