% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Catch_coverage.R
\name{MEDBS_Catch_coverage}
\alias{MEDBS_Catch_coverage}
\title{Check the coverage of Catch table}
\usage{
MEDBS_Catch_coverage(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{Catch table in MEDBS format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns two summary tables: one for landing coverage and the other for discard coverage. Furthermore, plots of landings and discards by gear are also returned
}
\description{
The function allows to check the coverage of Catch table by mean of summary tables summarizing both landing and discard volumes and producing relative plots for the selected species.
}
\examples{
MEDBS_Catch_coverage(Catch_tab_example, "DPS", "ITA", "GSA 9")
}
