% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Catch_LW.R
\name{MEDBS_Catch_LW}
\alias{MEDBS_Catch_LW}
\title{check the length-weight consistency in the Catch table}
\usage{
MEDBS_Catch_LW(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{Catch table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
the function returns a plot with the lengths and weights from the Catch table by year.
}
\description{
check the length-weight consistency in the Catch table
}
\examples{
MEDBS_Catch_LW(data = Catch_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Isabella Bitetto \href{mailto:bitetto@fondazionecoispa.org}{bitetto@fondazionecoispa.org}
}
