% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_check_duplicates.r
\name{MEDBS_check_duplicates}
\alias{MEDBS_check_duplicates}
\title{Check for duplicated data rows in landings, discards and catch tables}
\usage{
MEDBS_check_duplicates(data, type = "l", SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing either landings, discards or catch data}

\item{type}{string vector indicating the type of table to be checked. "l" for landing; "d" for discards; "c" for catch table.}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a data frame containing the duplicated rows to be likely deleted from the data.
}
\description{
The function checks the presence of duplicated rows in landings, discards or catch data.
}
\examples{
MEDBS_check_duplicates(data = Discard_tab_example, type = "d", SP = "DPS",
MS = "ITA", GSA = "GSA 9", verbose = TRUE)
MEDBS_check_duplicates(data = Landing_tab_example, type = "l", SP = "DPS",
MS = "ITA", GSA = "GSA 9", verbose = TRUE)
MEDBS_check_duplicates(data = Catch_tab_example, type = "c", SP = "DPS",
MS = "ITA", GSA = "GSA 9", verbose = TRUE)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
