% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_check_missing_years.r
\name{MEDBS_check_missing_years}
\alias{MEDBS_check_missing_years}
\title{Check for missing years in MED & BS tables}
\usage{
MEDBS_check_missing_years(data, type = "l", SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing either landings, discards or catch data}

\item{type}{string vector indicating the type of table to be checked. "l" for landing; "d" for discards; "c" for catch table; "gp" for growth parameters table; "alk" for age-length keys table; "ma" for maturity at age table; "ml" for maturity at length table; "sra" for sex ratio at age table; "srl" for maturity at length table..}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a vector containing the missing years in data. NULL is returned in case no data were available for the selected combination of country, GSA and species. A vector of length 0 is returned in case there are no missing years in the data.
}
\description{
The function checks the presence of missing years in MED & BS tables data.
}
\examples{
df <- Discard_tab_example[-which(Discard_tab_example$year==2011),]
MEDBS_check_missing_years(
  data = df, type = "d", SP = "DPS",
  MS = "ITA", GSA = "GSA 9", verbose = TRUE
)
}
\author{
Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
