% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_comp_disc_YQ.r
\name{MEDBS_comp_disc_YQ}
\alias{MEDBS_comp_disc_YQ}
\title{Comparison between discards in weight by quarter and -1}
\usage{
MEDBS_comp_disc_YQ(data, MS, GSA, SP, verbose = FALSE)
}
\arguments{
\item{data}{data frame containing discards data}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{SP}{species code}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
The function returns a data frame with the comparison of discards aggregated by quarters and by year
}
\description{
The function compares the discards weights aggregated by quarter and by year for a selected species at gear level.
}
\examples{
MEDBS_comp_disc_YQ(data = Discard_tab_example, MS = "ITA", GSA = "GSA 9", SP = "DPS")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
