% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_comp_disc_YQ_fishery.r
\name{MEDBS_comp_disc_YQ_fishery}
\alias{MEDBS_comp_disc_YQ_fishery}
\title{Comparison between discards in weight by quarter, quarter -1 and by fishery}
\usage{
MEDBS_comp_disc_YQ_fishery(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing discards data}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a data frame with the comparison of discards aggregated by quarters and by year and fishery
}
\description{
The function allow to estimates the discards in weight for a selected species by quarter and fishery
}
\examples{
MEDBS_comp_disc_YQ_fishery(data = Discard_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
