% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_comp_land_Q_VL.R
\name{MEDBS_comp_land_Q_VL}
\alias{MEDBS_comp_land_Q_VL}
\title{Comparison between landings in weight by quarter accounting for vessel length}
\usage{
MEDBS_comp_land_Q_VL(land, MS, GSA, SP)
}
\arguments{
\item{land}{data frame containing landing data}

\item{MS}{member state code as it is reported in the landing data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}
}
\value{
The function returns a dataframe for the comparison of landings aggregated by quarters accounting for the presence of vessel length information.
}
\description{
The function allows to perform the comparison of landings of a selected species aggregated by quarters accounting for the presence of vessel length
}
\examples{
MEDBS_comp_land_Q_VL(land=Landing_tab_example,MS="ITA",GSA=9,SP="DPS")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
