% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_disc_mean_weight.r
\name{MEDBS_disc_mean_weight}
\alias{MEDBS_disc_mean_weight}
\title{Check of mean discard weight aggregations by year, gear and fishery}
\usage{
MEDBS_disc_mean_weight(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing discards data}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a plot of the mean discards weight by year, gear and fishery aggregation
}
\description{
The function allows to check consistency of mean discard of a selected species plotting the discards' weight by year, gear and fishery
}
\examples{
MEDBS_disc_mean_weight(data = Discard_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
