% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Discard_coverage.R
\name{MEDBS_discard_coverage}
\alias{MEDBS_discard_coverage}
\title{Check the coverage of discard data}
\usage{
MEDBS_discard_coverage(Discard_tab, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{Discard_tab}{Discard table in MED&BS format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
A summary table and a plot of discard time series by year and gear are returned.
}
\description{
The function allows to check the coverage of the time series in discard table for a selected species.
}
\examples{
MEDBS_discard_coverage(Discard_tab_example, "DPS", "ITA", "GSA 9")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
