% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_GP_check.R
\name{MEDBS_GP_check}
\alias{MEDBS_GP_check}
\title{Check of growth parameters table}
\usage{
MEDBS_GP_check(data, SP, MS, GSA)
}
\arguments{
\item{data}{growth parameters table in MED&BS data call format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}
}
\value{
A list of objects containing a summary table and different plots of the growth curves by sex and year is returned by the function.
}
\description{
The function checks the growth parameters by sex and year for a selected species
}
\examples{
MEDBS_GP_check(GP_tab_example, "MUT", "ITA", "GSA 18")
}
