% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Landing_coverage.R
\name{MEDBS_Landing_coverage}
\alias{MEDBS_Landing_coverage}
\title{Check the coverage of Landing table}
\usage{
MEDBS_Landing_coverage(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{Landing table in MED&BS format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
A list containing a summary table and coverage plot is provided.
}
\description{
The function allows to check the coverage in landing table providing a summary table and a plot of landing.
}
\examples{
MEDBS_Landing_coverage(Landing_tab_example, "DPS", "ITA", "GSA 9")
}
