% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_length_ind.r
\name{MEDBS_length_ind}
\alias{MEDBS_length_ind}
\title{Main length size indicators}
\usage{
MEDBS_length_ind(
  data,
  type,
  SP,
  MS,
  GSA,
  splines = c(0.2, 0.4, 0.6, 0.8),
  Xtresholds = c(0.25, 0.5, 0.75),
  verbose = TRUE,
  embedded = FALSE
)
}
\arguments{
\item{data}{data frame of either landings or discards data}

\item{type}{type of data frame. "l" for landing and "d" for discard}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{splines}{spline values assignment to fit cumulative distributions}

\item{Xtresholds}{threshold value}

\item{verbose}{boolean. If TRUE messages are returned}

\item{embedded}{boolean. If TRUE, the function generates only Min, Median and Max
in a single combined plot. If FALSE, the combined plot includes also Mean, 25th and 75th percentiles.}
}
\value{
The function returns a list containing a summary table and a combined plot of length indicators.
}
\description{
The function allows to check the consistency of length data for a selected species
on both landings and discards: Main length size indicators. The function keeps the same
calculations and checks of the original implementation, but it returns a single combined
time series plot (instead of multiple plots).
}
\examples{
MEDBS_length_ind(Landing_tab_example,
  type = "l", SP = "DPS", MS = c("ITA"),
  GSA = c("GSA 9"), splines = c(0.2, 0.4, 0.6, 0.8),
  Xtresholds = c(0.25, 0.5, 0.75), embedded = FALSE
)
MEDBS_length_ind(Discard_tab_example,
  type = "d", SP = "DPS", MS = c("ITA"),
  GSA = c("GSA 9"), splines = c(0.2),
  Xtresholds = c(0.5), embedded = TRUE
)

}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
