% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_MA_check.R
\name{MEDBS_MA_check}
\alias{MEDBS_MA_check}
\title{Check of MA_tab (maturity at age) table}
\usage{
MEDBS_MA_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{maturity at AGE table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
A summary table and plots are returned by the function.
}
\description{
The function checks the maturity at age (MA) table providing a summary table of the data coverage and plots for the selected species of the proportion of matures for age class by sex and year.
}
\examples{
MEDBS_MA_check(MA_tab_example, "DPS", "ITA", "GSA 99")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
