% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_plot_disc_vol.r
\name{MEDBS_plot_disc_vol}
\alias{MEDBS_plot_disc_vol}
\title{Plot of total discards by gear and fishery}
\usage{
MEDBS_plot_disc_vol(data, SP, MS, GSA)
}
\arguments{
\item{data}{data frame containing discard data}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}
}
\value{
The function returns a plot of the total discards time series by fishery and gear
}
\description{
The function allows to visual check the time series of discard volumes by fishery of a selected species
}
\examples{
MEDBS_plot_disc_vol(data = Discard_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
