% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_SA_check.R
\name{MEDBS_SA_check}
\alias{MEDBS_SA_check}
\title{SA_tab (sex ratio at age) table check}
\usage{
MEDBS_SA_check(SA_tab, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{SA_tab}{sex ratio at age table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a summary table and plots
}
\description{
The function allows to check the sex ratio at age (SA) table providing a summary table of the data coverage and plots for the selected species of the proportion of sex ratio for age class by year.
}
\examples{
MEDBS_SA_check(SA_tab_example, "DPS", "ITA", "9")
}
