% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_SL_check.R
\name{MEDBS_SL_check}
\alias{MEDBS_SL_check}
\title{SL_tab (sex ratio at length) table check}
\usage{
MEDBS_SL_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{sex ratio at length table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a summary table and plots
}
\description{
The function allows to check the sex ratio at length (SL) table providing a summary table of the data coverage and plots for the selected species of the proportion of sex ratio for length class by year.
}
\examples{
MEDBS_SL_check(SL_tab_example, "DPS", "ITA", "GSA 99")
}
