% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_SOP.R
\name{MEDBS_SOP}
\alias{MEDBS_SOP}
\title{um of products}
\usage{
MEDBS_SOP(data, SP, MS, GSA, threshold = 5, verbose = TRUE)
}
\arguments{
\item{data}{Catch table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{threshold}{threshold value in percentage to check the percentage difference between SOP and both volumes of landing and discard. Default value is 5\%}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
the function returns a data frame of the record with values of percentage difference between SOP and both volumes of landing and discard greater then threshold value.
}
\description{
um of products
}
\examples{
MEDBS_SOP(data=Catch_tab_example,SP="DPS",MS="ITA",GSA="GSA 9",threshold = 5)
}
