% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_check.R
\name{ML_tab_check}
\alias{ML_tab_check}
\title{ML_tab (maturity at length) table check}
\usage{
ML_tab_check(ML_tab)
}
\arguments{
\item{ML_tab}{maturity at length table in MED&BS datacall format}
}
\value{
a summary table and plots
}
\description{
ML_tab (maturity at length) table check
}
\examples{
ML_tab_check(ML_tab_example)
}
