% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_tab_example.R
\docType{data}
\name{ML_tab_example}
\alias{ML_tab_example}
\title{ML table in MED&BS datacall format}
\format{
An object of class \code{data.frame} with 437 rows and 12 columns.
}
\usage{
ML_tab_example
}
\description{
ML table in MED&BS datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{MED&BS}
\keyword{datacall}
