% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_AL.R
\name{RCG_check_AL}
\alias{RCG_check_AL}
\title{Check consistency of age-length relationship}
\usage{
RCG_check_AL(data, MS, GSA, SP, min_age = 0, max_age = 30, verbose = TRUE)
}
\arguments{
\item{data}{table of detailed data in RCG format}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{SP}{species code}

\item{min_age}{minimum age expected}

\item{max_age}{maximum age expected}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a list of object containing a summary table, an age-length plot and a table of the outlayers, if any.
}
\description{
Check consistency of age-length relationship
}
\examples{
RCG_check_AL(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus", min_age = 0, max_age = 30)
}
