% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_CL.R
\name{RCG_check_CL}
\alias{RCG_check_CL}
\title{Quality checks on CL RCG table}
\usage{
RCG_check_CL(data, MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{Landing table in RCG CL format}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{SP}{species code}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The output is a list of 6 data frames:
\enumerate{
\item Sum of Landings by year, quarter and month;
\item Sum of Landing value by year, quarter and month;
\item Sum of landings by LandCtry, VslFlgCtry,  Area, Rect, SubRect, Harbour;
\item Sum of landing value by LandCtry, VslFlgCtry,  Area, Rect, SubRect, Harbour;
\item Sum of landings by Year, Species, foCatEu5, foCatEu6;
\item Sum of landing value by Year, Species, foCatEu5, foCatEu6.
}
}
\description{
The function checks CL tables for temporal, spatial, species and metier coverage
}
\examples{
RCG_check_CL(data_exampleCL, MS = "COUNTRY1", GSA = "GSA99", SP = "Parapenaeus longirostris")
}
\author{
Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
