% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_LFD.r
\name{RCG_check_LFD}
\alias{RCG_check_LFD}
\title{RCG_check LFD}
\usage{
RCG_check_LFD(data, MS, GSA, SP, min_len = 1, max_len = 1000, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{MS}{member state code as it is reported in the landing data}

\item{GSA}{string value of the GSA code}

\item{SP}{reference species for the analysis}

\item{min_len}{minimum length}

\item{max_len}{maximum length}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
comparison plot of LFDs among the  years and check of length range of the data using allowed range
}
\description{
The function allows to check the consistency of LFDs (length frequency distributions) by year on a given species generating a multi-frame plot. The function also returns the records in which the length classes are greater or lower than the expected ones (\code{min_len} and \code{max_len} parameters).
}
\examples{
RCG_check_LFD(data_ex,MS="ITA",GSA="GSA99", SP="Mullus barbatus",min_len=1,max_len=35)
}
