% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_LFD_comm_cat.R
\name{RCG_check_LFD_comm_cat}
\alias{RCG_check_LFD_comm_cat}
\title{Check consistency of LFD by year and commercial category}
\usage{
RCG_check_LFD_comm_cat(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a multi-frame plot of LFDs (length frequency distributions) by year and commercial size category for the selected species. The function also returns a data frame with the length range by year and commercial size category.
}
\description{
The function allows to check the consistency of LFDs (length frequency distributions) by year and commercial size category on a given species.
}
\examples{
RCG_check_LFD_comm_cat(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus")
}
