% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_lw.R
\name{RCG_check_lw}
\alias{RCG_check_lw}
\title{Consistency check of length-weight relationship}
\usage{
RCG_check_lw(data, SP, MS, GSA, Min = NA, Max = NA, verbose = TRUE)
}
\arguments{
\item{data}{table of detailed data in RCG format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{Min}{min weight expected in the data}

\item{Max}{max weight expected in the data}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
Plot and error message
}
\description{
The function allows to check the consistency of length-weight relationship by sex and year on a given species generating a multi-frame plot. The function also returns the records in which the individual weights are greater or lower than the expected ones (\code{Min} and \code{Max} parameters). In case the \code{Min} and \code{Max} parameters are not provided, the function tests the lowest and the higher values of individual weights as outliers by mean of the Grubbs' test (from package \code{outliers})
}
\examples{
RCG_check_lw(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus", Min = 0, Max = 1000)
RCG_check_lw(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus")
}
\references{
Grubbs, F.E. (1950). Sample Criteria for testing outlying observations. Ann. Math. Stat. 21, 1, 27-58
}
