% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_mat_ogive.R
\name{RCG_check_mat_ogive}
\alias{RCG_check_mat_ogive}
\title{check_mat_ogive}
\usage{
RCG_check_mat_ogive(
  data,
  MS,
  GSA,
  SP,
  sex,
  immature_stages = c("0", "1", "2a"),
  verbose = TRUE
)
}
\arguments{
\item{data}{table of detailed data in RCG format}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{SP}{species code}

\item{sex}{defines the sex of the individuals selected for the analysis ('F' for females, 'M' for males)}

\item{immature_stages}{vector of maturity stages considered immature}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
Plot ogive by sex
}
\description{
The function allows to check the maturity stages composition estimating the maturity ogives by sex for the selected species, using a selected pool of stages to classify the immature stages in the sample.
}
\examples{
RCG_check_mat_ogive(data_ex,
  MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus",
  sex = "F", immature_stages = c("0", "1", "2a")
)
}
