% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_summarize_ind_meas.R
\name{RCG_summarize_ind_meas}
\alias{RCG_summarize_ind_meas}
\title{Number of individual by trip for which biological data have been collected (length, sex, maturity, weight and age)}
\usage{
RCG_summarize_ind_meas(data, MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{detailed data in RCG CS format}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a list containing a summary of measurements by trip for each biological variable
}
\description{
Number of individual by trip for which biological data have been collected (length, sex, maturity, weight and age)
}
\examples{
RCG_summarize_ind_meas(data = data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus")
}
