% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_summarize_trips.R
\name{RCG_summarize_trips}
\alias{RCG_summarize_trips}
\title{summarizing the number of trips/hauls monitored by year by port, metier, sampling method;}
\usage{
RCG_summarize_trips(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
Number of trips by area, year, port, metier and sampling method
}
\description{
The function allows to summarise monitored by year by port, metier, sampling method
}
\examples{
RCG_summarize_trips(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus")
}
