% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.coded.file.r
\name{read.coded.file}
\alias{read.coded.file}
\title{Read encoded data}
\usage{
read.coded.file(file, sep = ",", dec = ".", verbose = FALSE)
}
\arguments{
\item{file}{file path of the table to be read}

\item{sep}{separator of the csv file (es. ";", ",")}

\item{dec}{decimals separator of the csv file (es. ".", ",")}

\item{verbose}{boolean. Returns messages if TRUE.}
}
\value{
The function returns the csv auto-selecting the best ecoding to be used. It is possible to set both the value and the decimar separators
}
\description{
Read encoded data
}
\examples{
## when fill = TRUE
test1 <- data.frame(a=1:5, b=c("a","b","c","d","e"))
tf <- tempfile()
write.table(test1, tf, sep=",",row.names=FALSE)
read.coded.file(tf,verbose=TRUE)
}
