% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_tab_check.R
\name{SA_tab_check}
\alias{SA_tab_check}
\title{SA_tab (sex ratio at age) table check}
\usage{
SA_tab_check(SA_tab, SP, MS, GSA)
}
\arguments{
\item{SA_tab}{sex ratio at age table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}
}
\value{
a summary table and plots
}
\description{
SA_tab (sex ratio at age) table check
}
\examples{
SA_tab_check(SA_tab_example,"DPS","ITA","9")
}
