% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.r
\docType{data}
\name{SPs}
\alias{SPs}
\title{SPs table}
\format{
An object of class \code{data.frame} with 13417 rows and 8 columns.
}
\usage{
SPs
}
\description{
SPs table
}
\author{
Walter Zupa \email{zupa@fondazionecoispa.org}
}
\keyword{SPs}
