% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ind_meas.R
\name{summarize_ind_meas}
\alias{summarize_ind_meas}
\title{Number of individual by trip for which biological data have been collected (length, sex, maturity, weight and age)}
\usage{
summarize_ind_meas(data_example)
}
\arguments{
\item{data_example}{detailed data in RCG CS format}
}
\value{
a list containing a summary of measurements by trip foreach biological variable
}
\description{
Number of individual by trip for which biological data have been collected (length, sex, maturity, weight and age)
}
\examples{
summarize_ind_meas(data_ex)
}
