% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_trips.R
\name{summarize_trips}
\alias{summarize_trips}
\title{summarizing the number of trips/hauls monitored by year by port, metier, sampling method;}
\usage{
summarize_trips(data_example)
}
\arguments{
\item{data_example}{RCG CS table}
}
\value{
Number of trips by area, year, port, metier and sampling method
}
\description{
summarizing the number of trips/hauls monitored by year by port, metier, sampling method;
}
\examples{
summarize_trips(data_ex)
}
