% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_empty_field_G.R
\name{check_EF_FDI_G}
\alias{check_EF_FDI_G}
\title{Check empty fields in FDI G table}
\usage{
check_EF_FDI_G(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task G table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
Two lists are returned by the function. The first list returns the number of NA for each reference column. The second list provides the index of each NA in the reference column.
}
\description{
The function checks the presence of not allowed empty data in the given table.
}
\examples{
check_EF_FDI_G(fdi_g_effort)
}
\author{
Loredana Casciaro \href{mailto:casciaro@coispa.eu}{casciaro@coispa.eu}

Sebastien Alfonso \href{mailto:salfonso@coispa.eu}{salfonso@coispa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
