% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_empty_field_J.R
\name{check_EF_FDI_J}
\alias{check_EF_FDI_J}
\title{Check empty fields in FDI J table}
\usage{
check_EF_FDI_J(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task J table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
two lists. The first list gives the number of NA for each column of interest. The second list gives the index of each NA in the column of interest
}
\description{
Check empty fields in FDI J table
}
\examples{
check_EF_FDI_J(fdi_j_capacity)

}
