% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_LFD.r
\name{check_LFD}
\alias{check_LFD}
\title{check LFD}
\usage{
check_LFD(data)
}
\arguments{
\item{data}{}
}
\value{
comparison plot of LFDs among the  years and length range of the data
}
\description{
check LFD
}
\examples{
check_LFD(dps9)

}
