% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_LFD_comm_cat.R
\name{check_LFD_comm_cat}
\alias{check_LFD_comm_cat}
\title{Title}
\usage{
check_LFD_comm_cat(data_example)
}
\arguments{
\item{data_example}{}
}
\value{
plot
}
\description{
Title
}
\examples{
check_LFD_comm_cat(data_example)
}
