% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_LFD_comm_cat.R
\name{check_LFD_comm_cat}
\alias{check_LFD_comm_cat}
\title{Check consistency of LFD by year and commercial category}
\usage{
check_LFD_comm_cat(data, species, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{species}{reference species for the analysis}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
plot and a summary table with ranges by year and commercial category
}
\description{
Check consistency of LFD by year and commercial category
}
\examples{
check_LFD_comm_cat(data_ex, species="Mullus barbatus")
}
