% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_loc.R
\name{check_loc}
\alias{check_loc}
\title{Check trip location}
\usage{
check_loc(data, ports = circabc)
}
\arguments{
\item{data}{detailed data in RCG CS format}

\item{ports}{ports codification file}
}
\value{
map of trip locations
}
\description{
The function allows to check the spatial distribution of data using the initial and final coordinates, where available, and the ports position in cluded in the data.

If Initial and/or Final coordinates are included in the data, maps of them are produced. If not the locations of the harbours are mapped.
}
\examples{
check_loc(data_ex)
}
