% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lw.R
\name{check_lw}
\alias{check_lw}
\title{Consistency of length-weight relationship and consistency with allowed ranges}
\usage{
check_lw(data_example, Min = 0, Max = 1000)
}
\arguments{
\item{data_example}{table of detailed data in RCG format}

\item{Min}{min weight expected in the data}

\item{Max}{max weight expected in the data}
}
\value{
Plot and error message
}
\description{
Consistency of length-weight relationship and consistency with allowed ranges
}
\examples{
check_lw(data_ex,Min=0,Max=1000)
}
