% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_relatioship_length_weight_VII2.R
\name{check_lw_TaskVII.2}
\alias{check_lw_TaskVII.2}
\title{Plot of the relationship length weight for each species}
\usage{
check_lw_TaskVII.2(data)
}
\arguments{
\item{data}{GFCM Task VII.2 table}
}
\value{
Plot of the length weight relationship for each species
}
\description{
Plot of the relationship length weight for each species
}
\examples{
check_lw_TaskVII.2(task_vii2)
}
