% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mat.R
\name{check_mat}
\alias{check_mat}
\title{Check consistency sex and maturity stage}
\usage{
check_mat(data, species, verbose = TRUE)
}
\arguments{
\item{data}{table of detailed data in RCG format}

\item{species}{reference species for the analysis}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
The function the returns the plot of the maturity stages by length class
}
\description{
The function allows to perform a visual check of the maturity stage composition by length class, sex and year.
}
\examples{
check_mat(data_ex,species="Mullus barbatus")
}
