% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_min_max_length_VII2.R
\name{check_minmaxl_TaskVII.2}
\alias{check_minmaxl_TaskVII.2}
\title{Comparison between min/max observed for each species with theoretical values}
\usage{
check_minmaxl_TaskVII.2(data, tab_length, MS, GSA)
}
\arguments{
\item{data}{GFCM Task II.2 table}

\item{tab_length}{Theoretical values of min/max length for each species}

\item{MS}{member state code}

\item{GSA}{GSA code}
}
\value{
The function returns a table with the comparison between min/max lengths observed for each species with theoretical values.
}
\description{
Function to verify the consistency of the lengths reported in the TaskVII.2 table with the theoretical values reported in the minmaxLtaskVII2 table. The function allows to identify the records in which the observed lengths are greater or lower than the expected ones.
}
\examples{
check_minmaxl_TaskVII.2(task_vii2,minmaxLtaskVII2,MS="ITA",GSA="18")
}
