% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_Check_record_duplicated_J.R
\name{check_RD_FDI_J}
\alias{check_RD_FDI_J}
\title{Check duplicated records in FDI J table}
\usage{
check_RD_FDI_J(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task J table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 7 columns
}
\description{
Check duplicated records in FDI J table
}
\examples{
check_RD_FDI_J(fdi_j_capacity)
}
