% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFCM_Check_record_duplicated_VII31.R
\name{check_RD_taskVII31}
\alias{check_RD_taskVII31}
\title{Check duplicated records in GFCM Task VII.3.1 table}
\usage{
check_RD_taskVII31(data, verbose = TRUE)
}
\arguments{
\item{data}{GFCM Task VII.3.1 table}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
indices of the duplicated rows in the first 5 columns
}
\description{
Check duplicated records in GFCM Task VII.3.1 table
}
\examples{
check_RD_taskVII31(task_vii31)
}
