% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDI_AER_land_landvalue.r
\name{FDI_AER_land_landvalue}
\alias{FDI_AER_land_landvalue}
\title{Cross check of landings or landings' value between the FDI and AER databases}
\usage{
FDI_AER_land_landvalue(
  FDI,
  AER,
  var = "landings",
  MS,
  level = "GSA",
  YEAR = NA,
  GSA = NA,
  SP = NA,
  OUT = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{FDI}{(mandatory) data frame containing FDI catch data (i.e. TABLE_A_CATCH)}

\item{AER}{(mandatory) data frame containing AER catch data (i.e. map_fsfao)}

\item{var}{(mandatory) string to select the variable of interest. Use either 'landings' for landings
or 'value' for landings' value.}

\item{MS}{(mandatory) The MS code (e.g. 'GRC'). Only one MS code is allowed.}

\item{level}{(optional) character to select the level of aggregation of variable 'var', i.e.
'MS', 'GSA', 'GEAR' or 'METIER'. E.g. if level 'GEAR' is selected, the comparison
will be performed by 'MS', 'GSA' and 'GEAR'. Default level is 'METIER'.}

\item{YEAR}{(optional) vector of years to perform the check on. Default is NA, which will produce a check for all
the years found in both data frames}

\item{GSA}{(optional) vector string of GSAs for the given MS to produce the check on (e.g. GSA = c('GSA20', 'GSA22')).
Default is NA, which will produce the check for all GSAs existing in the data frames.}

\item{SP}{(optional) vector string of species FAO 3-alpha codes to produce the checks on (e.g. SPECIES = c('DPS', 'TGS')).
Default is NA, which will produce the check for all SP existing in the data frames.}

\item{OUT}{If set as TRUE tables in csv will be saved in the OUTPUT folder created in the working directory. Default is FALSE}

\item{verbose}{(optional) boolean, if TRUE messages are printed. Default is FALSE.}
}
\value{
The function returns a list of three data frames, containing: a) entries for which 'var' differs between
the two data frames (mismatch_FDI_AER) reporting only differences greater than 5\%, b) entries in AER that are missing from FDI (missing_FDI) and
c) entries in FDI that are missing from AER (missing_AER). This output is also saved in .csv files if OUT is TRUE.
}
\description{
The function compares landings (var='landings') or landings' value (var='value) by SP and YEAR between
the FDI and AER databases for the given MS. By default, the comparison is done by GSA, GEAR and METIER.
Optionally, the user can select a higher level of aggregation (e.g. 'MS', 'GSA', etc.) (see level variable)
}
\examples{
FDI_AER_land_landvalue(FDI=fdi_a_catch, AER=aer_catch, var='landings',
level='GEAR', MS='PSP', YEAR=NA, GSA=c('GSA97', 'GSA98'), SP=NA, OUT = TRUE, verbose = TRUE)
}
\author{
Vasiliki Sgardeli \href{mailto:vsgard@hcmr.gr}{vsgard@hcmr.gr}
}
