% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_tab_check.R
\name{MA_tab_check}
\alias{MA_tab_check}
\title{MA_tab (maturity at age) table check}
\usage{
MA_tab_check(MA_tab)
}
\arguments{
\item{MA_tab}{maturity at AGE table in MED&BS datacall format}
}
\value{
a summary table and plots
}
\description{
MA_tab (maturity at age) table check
}
\examples{
MA_tab_check(MA_tab_example)
}
