% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Catch_CAA.R
\name{MEDBS_Catch_CAA}
\alias{MEDBS_Catch_CAA}
\title{Check the numbers-at-age in the Catch table}
\usage{
MEDBS_Catch_CAA(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{Catch table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
A ggplot object: multi-panel barplots of numbers-at-age (quarters summed),
by YEAR (columns) and by group GEAR - MESH_SIZE_RANGE - FISHERY (rows).
}
\description{
Check the numbers-at-age in the Catch table
}
\examples{
MEDBS_Catch_CAA(data = Catch_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Isabella Bitetto \href{mailto:bitetto@fondazionecoispa.org}{bitetto@fondazionecoispa.org}
}
