% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_comp_disc_YQ_fishery.r
\name{MEDBS_comp_disc_YQ_fishery}
\alias{MEDBS_comp_disc_YQ_fishery}
\title{Comparison between discards in weight by quarter, quarter -1 and by fishery}
\usage{
MEDBS_comp_disc_YQ_fishery(disc, MS, GSA, SP)
}
\arguments{
\item{disc}{data frame containing discards data}

\item{MS}{member state code as it is reported in the discards data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}
}
\value{
The function returns a dataframe  for the comparison of discards aggregated by quarters and by year and fishery
}
\description{
Comparison between discards in weight by quarter, quarter -1 and by fishery
}
\examples{
MEDBS_comp_disc_YQ(disc=discards,MS="ITA",GSA=18,SP="MUT")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
