% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_disc_mean_weight.r
\name{MEDBS_disc_mean_weight}
\alias{MEDBS_disc_mean_weight}
\title{Mean weight by year,gear and fishery aggregation}
\usage{
MEDBS_disc_mean_weight(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing discards data}

\item{SP}{species reference code in the three alpha code format}

\item{MS}{member state code as it is reported in the discards data}

\item{GSA}{GSA code}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
The function returns a plot of the mean discards weight by year, gear and fishery aggregation
}
\description{
The function allows to check consistency of  mean discard of a selected species plotting the discards' weight by year, gear and fishery
}
\examples{
MEDBS_disc_mean_weight(data = Discard_tab_example, SP = "DPS", MS = "ITA", GSA = "GSA 9")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
