% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Discard_coverage.R
\name{MEDBS_discard_coverage}
\alias{MEDBS_discard_coverage}
\title{Discard_cov: function to check the coverage in Discard table}
\usage{
MEDBS_discard_coverage(Discard_tab, SP, MS, GSA)
}
\arguments{
\item{Discard_tab}{Discard table in MED&BS format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}
}
\value{
summary table and plots
}
\description{
Discard_cov: function to check the coverage in Discard table
}
\examples{
MEDBS_discard_coverage(Discard_tab_example,"DPS","ITA","9")
}
