% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_ks.r
\name{MEDBS_ks}
\alias{MEDBS_ks}
\title{Kolmogorov-Smirnov test}
\usage{
MEDBS_ks(data, type, SP, MS, GSA, Rt = 1, verbose = TRUE)
}
\arguments{
\item{data}{data frame of either landings or discards data}

\item{type}{type of data frame. "l" for landing and "d" for discard}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{Rt}{ratio to be applied to subsample data to reduce the risk of rejection of H0 Hypothesis}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
the function returns a list of data frames and cumulative distribution plots
}
\description{
The function allows to perform the Kolmogorov-Smirnov test on both landings and discards for a selected species providing cumulative length distribution plots by fishery and year. The function performs Kolmogorov-Smirnov tests on couples of years to assess if they belong to the same population.
}
\examples{
MEDBS_ks(data = Landing_tab_example, type = "l", SP = "DPS", MS = "ITA", GSA = "GSA 9", Rt = 1)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
