% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_Landing_coverage.R
\name{MEDBS_Landing_coverage}
\alias{MEDBS_Landing_coverage}
\title{Landing_cov: function to check the coverage in Landing table}
\usage{
MEDBS_Landing_coverage(Landing_tab, SP, MS, GSA)
}
\arguments{
\item{Landing_tab}{Landing table in MED&BS format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}
}
\value{
summary table and plots
}
\description{
Landing_cov: function to check the coverage in Landing table
}
\examples{
MEDBS_Landing_coverage(Landing_tab_example,"DPS","ITA","9")
}
