% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_lengthclass_0.R
\name{MEDBS_lengthclass_0}
\alias{MEDBS_lengthclass_0}
\title{Length classes number 0 in landings and discards}
\usage{
MEDBS_lengthclass_0(data, type = "l", MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing landing data}

\item{type}{string vector indicating the type of table to be checked. "l" for landing; "d" for discards.}

\item{MS}{member state code as it is reported in the landing data}

\item{GSA}{string value of the GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{verbose}{Boolean value to obtain further explanation messages from the function}
}
\value{
The function returns a data frame with the rows with 0 values length class having weigth > 0.
}
\description{
The function checks landings and discards for the presence of length class filled in having weigth > 0.
}
\examples{
MEDBS_lengthclass_0(data=landing,type="l",MS="ITA",GSA=11,SP="ARA",verbose=TRUE)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
