% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_lengthclass_0.R
\name{MEDBS_lengthclass_0}
\alias{MEDBS_lengthclass_0}
\title{Checks length classes numbers with zeros in landings and discards}
\usage{
MEDBS_lengthclass_0(data, type = "l", SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data frame containing landing  or discards data}

\item{type}{string vector indicating the type of table to be checked. "l" for landing; "d" for discards.}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a data frame with the rows with 0 values length class having weigth > 0.
}
\description{
The function checks landings and discards for the presence of length class filled in having weigth > 0.
}
\examples{
MEDBS_lengthclass_0(
  data = Landing_tab_example, type = "l",
  SP = "DPS", MS = "ITA", GSA = "GSA 9", verbose = TRUE
)
MEDBS_lengthclass_0(
  data = Discard_tab_example, type = "d", SP = "DPS",
  MS = "ITA", GSA = "GSA 9", verbose = TRUE
)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
