% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_LW_check.R
\name{MEDBS_LW_check}
\alias{MEDBS_LW_check}
\title{Check LW parameters in GP table}
\usage{
MEDBS_LW_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{growth params table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
A summary table and plots of the LW parameters are returned by the function.
}
\description{
The function checks the length-weight parameters included in the GP table for a selected species.
}
\examples{
MEDBS_LW_check(GP_tab_example, "MUT", "ITA", "GSA 18")
}
\author{
Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
