% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_MA_check.R
\name{MEDBS_MA_check}
\alias{MEDBS_MA_check}
\title{MA_tab (maturity at age) table check}
\usage{
MEDBS_MA_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{maturity at AGE table in MED&BS datacall format}

\item{SP}{species (three alpha code)}

\item{MS}{Country}

\item{GSA}{GSA (Geographical sub-area (GFCM sensu))}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
a summary table and plots
}
\description{
The function allows to check the maturity at age (MA) table providing a summary table of the data coverage and plots for the selected species of the proportion of matures for age class by sex and year.
}
\examples{
MEDBS_MA_check(MA_tab_example,"DPS","ITA","GSA 99")
}
