% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_plot_discard_ts.R
\name{MEDBS_plot_discard_ts}
\alias{MEDBS_plot_discard_ts}
\title{Plot of total discards}
\usage{
MEDBS_plot_discard_ts(disc, MS, GSA, SP, by = "year")
}
\arguments{
\item{disc}{data frame containing discard data}

\item{MS}{member state code as it is reported in the discard data}

\item{GSA}{GSA code}

\item{SP}{species reference code in the three alpha code format}

\item{by}{string defining the temporal aggregation level of discard data to be plotted. Allowed values are: "year" and "quarter}
}
\value{
The function returns a plot of the total discard time series by year or by quarters. The plot by year also reports the landing by gear.
}
\description{
The function estimates the total discard time series by both year and quarters for a selected combination of member state, GSA and species.
}
\examples{
MEDBS_plot_discard_ts(disc=Discard_tab_example,MS="ITA",GSA=9,SP="DPS",by="quarter")
MEDBS_plot_discard_ts(disc=Discard_tab_example,MS="ITA",GSA=9,SP="DPS",by="year")
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
