% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_SA_check.R
\name{MEDBS_SA_check}
\alias{MEDBS_SA_check}
\title{Check of SA_tab (sex ratio at age) table}
\usage{
MEDBS_SA_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{sex ratio at age table in MED&BS datacall format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
a summary table and plots
}
\description{
The function allows to check the sex ratio at age (SA) table providing a summary table of the data coverage and plots for the selected species of the proportion of sex ratio for age class by year.
}
\examples{
MEDBS_SA_check(SA_tab_example, "DPS", "ITA", "GSA 99")
}
