% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_SL_check.R
\name{MEDBS_SL_check}
\alias{MEDBS_SL_check}
\title{Check of SL_tab (sex ratio at length) table}
\usage{
MEDBS_SL_check(data, SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{sex ratio at length table in MED&BS data call format}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
a summary table and plots
}
\description{
The function allows to check the sex ratio at length (SL) table providing a summary table of the data coverage and plots for the selected species of the proportion of sex ratio for length class by year.
}
\examples{
MEDBS_SL_check(SL_tab_example, "DPS", "ITA", "GSA 99")
}
\author{
Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}
}
