% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDBS_weight_0.R
\name{MEDBS_weight_0}
\alias{MEDBS_weight_0}
\title{weight 0 in landings and discards}
\usage{
MEDBS_weight_0(data, type = "l", SP, MS, GSA, verbose = TRUE)
}
\arguments{
\item{data}{data.table object containing landing or discard data}

\item{type}{type of table: "l" for landings; "d" for discards}

\item{SP}{species reference code in the three alpha code format}

\item{MS}{member state code as it is reported in both landing and discard data}

\item{GSA}{GSA code}

\item{verbose}{boolean value to obtain further explanation messages from the function}
}
\value{
The function returns the number of rows with 0 values in weights having length classes filled in.
}
\description{
The function checks landings or discards in weight equal to 0 having length classes filled in
}
\examples{
MEDBS_weight_0(data = Landing_tab_example, type = "l",
SP = "DPS", MS = "ITA", GSA = "GSA 9", verbose = TRUE)
MEDBS_weight_0(data = Discard_tab_example, type = "d", SP = "DPS",
MS = "ITA", GSA = "GSA 9", verbose = TRUE)
}
\author{
Alessandro Mannini \href{mailto:alessandro.mannini@ec.europa.eu}{alessandro.mannini@ec.europa.eu}

Walter Zupa \href{mailto:zupa@coispa.it}{zupa@coispa.it}

Isabella Bitetto \href{mailto:bitetto@coispa.it}{bitetto@coispa.it}
}
