% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_AL.R
\name{RCG_check_AL}
\alias{RCG_check_AL}
\title{Check consistency of age-length relationship}
\usage{
RCG_check_AL(data, MS, GSA, SP, min_age = 0, max_age = 30, verbose = TRUE)
}
\arguments{
\item{data}{table of detailed data in RCG format}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{reference species for the analysis}

\item{min_age}{minimum age expected}

\item{max_age}{maximum age expected}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
summary table length-age and error (if any)
}
\description{
Check consistency of age-length relationship
}
\examples{
RCG_check_AL(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus", min_age = 0, max_age = 30)
}
