% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_LFD.r
\name{RCG_check_LFD}
\alias{RCG_check_LFD}
\title{Consistency check of LFDs}
\usage{
RCG_check_LFD(data, SP, MS, GSA, min_len = NA, max_len = NA, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{SP}{species code}

\item{MS}{member state code}

\item{GSA}{GSA code (Geographical sub-area)}

\item{min_len}{minimum length}

\item{max_len}{maximum length}

\item{verbose}{boolean. If TRUE messages are returned}
}
\value{
The function returns a comparison plot of LFDs among the years and a table reporting the length classes out of the expected range, according to the reference length provided.
}
\description{
The function allows to check the consistency of LFDs (length frequency distributions) by year on a given species generating a multi-frame plot. The function also returns the records in which the length classes are greater or lower than the expected ones (\code{min_len} and \code{max_len} parameters).
}
\examples{
RCG_check_LFD(data=data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus",
min_len = 30, max_len = 300)
RCG_check_LFD(data=data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus",
min_len = NA, max_len = NA)
}
