% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCG_check_LFD_comm_cat.R
\name{RCG_check_LFD_comm_cat}
\alias{RCG_check_LFD_comm_cat}
\title{Check consistency of LFD by year and commercial category}
\usage{
RCG_check_LFD_comm_cat(data, MS, GSA, SP, verbose = TRUE)
}
\arguments{
\item{data}{RCG CS table}

\item{MS}{member state code}

\item{GSA}{GSA code}

\item{SP}{reference species for the analysis}

\item{verbose}{boolean. If it is TRUE messages are reported with the outputs}
}
\value{
plot and a summary table with ranges by year and commercial category
}
\description{
The function allows to check the consistency of LFDs (length frequency distributions) by year and commercial size cetegory on a given species generating a multi-frame plot. The function also returns a data frame with the length range by year and commercial size category.
}
\examples{
RCG_check_LFD_comm_cat(data_ex, MS = "ITA", GSA = "GSA99", SP = "Mullus barbatus")
}
