% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_ii2.R
\docType{data}
\name{task_ii2}
\alias{task_ii2}
\title{task_ii2 table in GFCM DCRF datacall format}
\format{
An object of class \code{data.frame} with 15 rows and 9 columns.
}
\usage{
task_ii2
}
\description{
task_ii2 table in GFCM DCRF datacall format
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{DCRF}
\keyword{GFCM}
\keyword{datacall}
